[FEAT] 소소피드 작성 API 구현

+++ src/main/java/org/websoso/WSSServer/domain/Novel.java
@@ -48,7 +48,7 @@ public class Novel {
     private Flag isCompleted;
 
     @Column(columnDefinition = "int default 0", nullable = false)
-    private Long novelRatingCount;
+    private Integer novelRatingCount;
 
     @Column(columnDefinition = "float default 0.0", nullable = false)
     private Float novelRatingSum;
